Add text selection util and update Header component to use it
